requirements.txt
pandas
numpy
pillow
colorama